integrated permissions

+++ Absoft/Controllers/DonViTinhController.cs
@@ -1,39 +1,38 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Absoft.Data;
-using Absoft.Data.Entities;
+﻿using Absoft.Authorization;
 using Absoft.Extentions;
 using Absoft.Helpers;
 using Absoft.Repositories.Interfaces;
 using Absoft.ViewModels;
-using AutoMapper;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
-using OfficeOpenXml;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace Absoft.Controllers
 {
     public class DonViTinhController : BaseController
     {
         IDonViTinhRepository _donViTinhRepository;
-        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IAuthorizationService _authorizationService;
 
         public List<int> JsonConnvert { get; private set; }
 
-        public DonViTinhController(IDonViTinhRepository donViTinhRepository, IHostingEnvironment hostingEnvironment)
+        public DonViTinhController(
+            IDonViTinhRepository donViTinhRepository,
+            IAuthorizationService authorizationService)
         {
             _donViTinhRepository = donViTinhRepository;
-            _hostingEnvironment = hostingEnvironment;
+            _authorizationService = authorizationService;
 
         }
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Read);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var models = await _donViTinhRepository.GetAllAsync();
             return Ok(models);
         }
@@ -41,6 +40,10 @@ namespace Absoft.Controllers
         [HttpGet("getAllPaging")]
         public async Task<IActionResult> GetAllPaging([FromQuery]PagingParams pagingParams)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Read);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var paged = await _donViTinhRepository.GetAllPagingAsync(pagingParams);
             Response.AddPagination(paged.CurrentPage, paged.PageSize, paged.TotalCount, paged.TotalPages);
             return Ok(paged.Items);
@@ -49,18 +52,30 @@ namespace Absoft.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetDetail(int id)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Read);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var models = await _donViTinhRepository.GetById(id);
             return Ok(models);
         }
         [HttpPut]
         public async Task<IActionResult> Update(DonViTinhViewModel donViTinhViewModel)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Update);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var result = await _donViTinhRepository.UpdateAsync(donViTinhViewModel);
             return Ok(result);
         }
         [HttpPost]
         public async Task<IActionResult> Insert(DonViTinhViewModel donViTinhViewModel)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Create);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var id = await _donViTinhRepository.CheckTonTai(donViTinhViewModel.TenDVT);
             if (id == -1)
             {
@@ -75,29 +90,23 @@ namespace Absoft.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Delete);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var result = await _donViTinhRepository.DeleteAsync(id);
             return Ok(result);
         }
         [HttpDelete("DeleteAllAsync/{strIds}")]
         public async Task<IActionResult> DeleteAllAsync(string strIds)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "DON_VI_TINH", Operations.Delete);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             // strids [1,2,3,4,4,5,56,6]
             List<int> listId = JsonConvert.DeserializeObject<List<int>>(strIds);
             var result = await _donViTinhRepository.DeleteAllAsync(listId);
-            return Ok(result);
-        }
-        [HttpPost]
-        [Route("ImportDVT")]
-        public async Task<IActionResult> ImportDVT(IList<IFormFile> files)
-        {
-            var result = await _donViTinhRepository.ImportDVT(files);
-            return Ok(result);
-        }
-        [HttpGet]
-        [Route("ExportDVT")]
-        public IActionResult ExportDVT()
-        {
-            var result = _donViTinhRepository.ExportDVT();
             return Ok(result);
         }
     }
